Minimal conversion of Javascript runtime to use GcNode instead of Rc
This conversion is necessary to support a mark-and-sweep GC for the runtime where
Rust's Rc will leave many cyclical references behind. This is NOT fully implemented
and garbage collection is expected to prune everything at the moment. This fixe type
issues and sets up the codebase to mostly work as expected